+++ config/application.rb
@@ -28,6 +28,11 @@ module Slice
     # Run "rails time:zones" for a list of tasks for finding time zone names. Default is UTC.
     config.time_zone = 'Eastern Time (US & Canada)'
 
+    # Overwrite Rails errors to use Bootstrap CSS classes
+    config.action_view.field_error_proc = Proc.new do |html_tag, instance|
+      "<span class=\"has-error\">#{html_tag}</span>".html_safe
+    end
+
     # Add RandomizationAlgorithm Module to autoload path
     config.autoload_paths << Rails.root.join('app', 'models', 'randomization_algorithm')
     config.autoload_paths << Rails.root.join('app', 'models', 'randomization_algorithm', 'algorithms')